% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_biocViews.R
\name{add_biocViews}
\alias{add_biocViews}
\title{Add biocViews Field to Description File}
\usage{
add_biocViews(value = "", overwrite = TRUE)
}
\arguments{
\item{value}{package names}

\item{overwrite}{logical, TRUE is defaulted}
}
\description{
Add biocViews Field to Description File
}
