% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.rules}
\alias{as.data.frame.calibrate}
\title{Transform to dataframe
rules object or calibrate object}
\usage{
\method{as.data.frame}{rules}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{calibrate}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{data with rules class for package 'arules'}

\item{row.names}{ignore}

\item{optional}{ignore}

\item{...}{ignore}
}
\value{
a dataframe
}
\description{
Transform to dataframe
rules object or calibrate object
}
