% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.rules.R
\name{as.data.frame.rules}
\alias{as.data.frame.rules}
\title{Transform rules to dataframe}
\usage{
\method{as.data.frame}{rules}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{data with rules class for package 'arules'}

\item{row.names}{ignore}

\item{optional}{ignore}

\item{...}{ignore}
}
\value{
a dataframe
}
\description{
Transform rules to dataframe
}
