% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.transactions.R
\name{as.transactions}
\alias{as.transactions}
\title{Transform to transactions}
\usage{
as.transactions(x)
}
\arguments{
\item{x}{dataframe or matrix}
}
\value{
a transaction data
}
\description{
Transform to transactions
}
