% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_href.R
\name{attr_href}
\alias{attr_href}
\title{Get hypertext reference attributes}
\usage{
attr_href(x)
}
\arguments{
\item{x}{A document (from read_html()), node set (from html_elements()), node (from html_element()), or session (from session()).}
}
\value{
hypertext reference attributes
}
\description{
Get hypertext reference attributes
}
