% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.xml_nodeset.R
\name{c.xml_nodeset}
\alias{c.xml_nodeset}
\title{Comine xml_nodeset}
\usage{
\method{c}{xml_nodeset}(...)
}
\arguments{
\item{...}{one or more xml_nodeset}
}
\value{
xml_nodeset
}
\description{
Comine xml_nodeset
}
