% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_n.R
\name{cat_n}
\alias{cat_n}
\title{print vector by lines}
\usage{
cat_n(x, n = 3, ind = 0)
}
\arguments{
\item{x}{one vector}

\item{n}{number if element in each line, default is 3}

\item{ind}{indentation, default is 0}
}
\value{
print vector by lines
}
\description{
print vector by lines
}
\examples{
cat_n(1:10)
cat_n(1:10,ind=3)
}
