% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{character.nms}
\alias{character.nms}
\title{Return character names in matrix or dataframe}
\usage{
character.nms(df)
}
\arguments{
\item{df}{dataframe or matrix}
}
\value{
character names vectors
}
\description{
Return character names in matrix or dataframe
}
