% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese_utf8.R
\name{chinese_utf8}
\alias{chinese_utf8}
\title{UTF8 Code for Chinese}
\usage{
chinese_utf8(x)
}
\arguments{
\item{x}{chinese characters}
}
\value{
an expression with UTF8 code.
}
\description{
UTF8 Code for Chinese
}
