% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGI_download_dir.R, R/CIGI_download_file.R,
%   R/CIGI_download_url.R
\name{CIGI}
\alias{CIGI}
\alias{CIGI_download}
\alias{CIGI_download_file}
\alias{CIGI_download_url}
\title{Download data from CIGI database}
\usage{
CIGI_download(url)

CIGI_download_file(url)

CIGI_download_url(url)
}
\arguments{
\item{url}{url link of data}
}
\value{
downloaded data
}
\description{
Download data from CIGI database
}
\examples{
\donttest{
CIGI_download('https://cigi-data.nci.nih.gov/Public/')
}
\donttest{
# download file by url
CIGI_download_file(url='https://cgci-data.nci.nih.gov/Public/BLGSP/biospecimen/BCCA/CGCI-BLGSP_BiospecimenSupplementSpreadsheet001_20191011.xlsx')
}
\donttest{
# get all downloaded urls
CIGI_download_url(url='https://cigi-data.nci.nih.gov/Public/')
}
}
