% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnOS.R
\name{cnOS}
\alias{cnOS}
\title{Chinese operating system
Whether the computer is Chinese operating system}
\usage{
cnOS()
}
\value{
logical
}
\description{
Chinese operating system
Whether the computer is Chinese operating system
}
\examples{
cnOS()
}
