% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_split.R
\name{col_split}
\alias{col_split}
\title{Split A Vector into Columns}
\usage{
col_split(x, split, reg_expr, colnames, cat = TRUE)
}
\arguments{
\item{x}{a vector}

\item{split}{one or more characters. Split exactly}

\item{reg_expr}{character. Split by regular expressions}

\item{colnames}{optional. Column names for outcome}

\item{cat}{logical, whether to show message}
}
\value{
A dataframe with several columns.
}
\description{
Split A Vector into Columns
}
\examples{
x=c('1a2','3a4','4a4')
col_split(x,split='a')
col_split(x = x,reg_expr = '[a-z]')

#two splits
df=data.frame(result=c('A, B-C',
                       'A, C-D',
                       'E, F-G'))
col_split(x = df[,1],split = c(',','-'))
}
