% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.it.numeric.it.R
\name{columntrans}
\alias{columntrans}
\alias{factor.it<-}
\alias{factor.it}
\alias{numeric.it}
\alias{numeric.it<-}
\title{Change data type}
\usage{
factor.it(x) <- value

factor.it(x, value)

numeric.it(x, value)

numeric.it(x) <- value
}
\arguments{
\item{x}{dataframe}

\item{value}{column names}
}
\value{
factor or numeric columns in a dataframe
}
\description{
Change data type
}
\examples{
str(mtcars)
factor.it(mtcars,c("cyl", "vs", "am", "gear"))
factor.it(mtcars)=c("cyl", "vs", "am", "gear")
str(mtcars)

numeric.it(mtcars,c("cyl", "vs", "am", "gear"))
numeric.it(mtcars)=c("cyl", "vs", "am", "gear")
str(mtcars)
}
