% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.data.R
\name{complete.data}
\alias{complete.data}
\title{Complete data}
\usage{
complete.data(x)
}
\arguments{
\item{x}{dataframe or matrix or atomic}
}
\value{
complete data
}
\description{
Removing rows with NA in dataframe or matrix. Removing NA atomic.
}
\examples{
x=c(1,NA,2)
complete.data(x)

x=data.frame(a=c(1,NA))
complete.data(x)
}
