% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTD2_download_dir.R, R/CTD2_download_file.R,
%   R/CTD2_download_url.R
\name{CTD2}
\alias{CTD2}
\alias{CTD2_download_dir}
\alias{CTD2_download_file}
\alias{CTD2_download_url}
\title{Download data from CTD2 database}
\usage{
CTD2_download_dir(url)

CTD2_download_file(url)

CTD2_download_url(url)
}
\arguments{
\item{url}{url link of data}
}
\value{
downloaded data
}
\description{
Download data from CTD2 database
}
\examples{
\donttest{
CTD2_download_dir('https://ctd2-data.nci.nih.gov/Public/')
}
\donttest{
# download files by file url
CTD2_download_file(url='https://ctd2-data.nci.nih.gov/Public/Broad/CTRPv1.0_2013_pub_Cell_154_1151/CTRPv1.0._COLUMNS.xlsx')
}
\donttest{
# get file url
CTD2_download_url(url='https://ctd2-data.nci.nih.gov/Public/Broad/CTRPv1.0_2013_pub_Cell_154_1151/')
}
}
