% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{current_mirror}
\alias{current_mirror}
\title{Current mirror}
\usage{
current_mirror()
}
\value{
a list contains CRAN and Bioconductor mirror
}
\description{
Current mirrors of CRAN adn Bioconductor
}
\examples{
current_mirror()
}
