% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_decrease.R
\name{decrease}
\alias{decrease}
\title{decrease character}
\usage{
decrease(chr)
}
\arguments{
\item{chr}{one character vector}
}
\value{
decreased vector
}
\description{
decrease character
}
\examples{
set.seed(2020)
x=rnorm(20)
decrease(x)
}
