% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_move.R
\name{delete_left}
\alias{delete_left}
\title{Delete and Move Left the rest Values}
\usage{
delete_left(x, delete)
}
\arguments{
\item{x}{dataframe or matrix}

\item{delete}{one delete object}
}
\value{
dataframe or matrix
}
\description{
Delete and Move Left the rest Values
}
\examples{
a=c(1,NA,7,NA)
b=c(NA,2,2,7)
d=c(1,NA,40,7)
df=data.frame(a,b,d)
delete_left(x=df,NA)
}
