% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse0.R
\name{deparse0}
\alias{deparse0}
\title{substitue, deparse and paste}
\usage{
deparse0(x)
}
\arguments{
\item{x}{one object}
}
\value{
character
}
\description{
substitue, deparse and paste
}
\examples{
\donttest{
deparse0(j)
}

}
