% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2df.R
\name{desc2df}
\alias{desc2df}
\title{Convert package description file to dataframe}
\usage{
desc2df(desc)
}
\arguments{
\item{desc}{description file path}
}
\value{
One dataframe with column names of field
}
\description{
Convert package description file to dataframe
}
