% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_as_tbl}
\alias{do_as_tbl}
\title{Convert do_* output to tibble}
\usage{
do_as_tbl(x)
}
\arguments{
\item{x}{Output from do_* function}
}
\value{
Print out of summary info and a tibble of the data.
}
\description{
Convert do_* response objects into informative tibbles
}
