% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet.R
\name{do_droplot}
\alias{do_droplot}
\title{Create droplot}
\usage{
do_droplot(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to convert response into a tibble.
Defaults to yes.}
}
\value{
Either a response object or a tibble
}
\description{
Create a new droplet
}
