% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.it.R
\name{dump.it}
\alias{dump.it}
\title{Create dump matrix for a vector}
\usage{
dump.it(..., include.name = TRUE)
}
\arguments{
\item{...}{one vector}

\item{include.name}{logical, default is TRUE, wether to include name of variable}
}
\value{
a dump matix contains 0 and 1
}
\description{
Create dump matrix for a vector
}
\examples{
x=c('a','b','c','a','a')
dump.it(x)
dump.it(mtcars$am)
dump.it(mtcars[,'am'])
}
