% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup.connect.R
\name{dup.connect}
\alias{dup.connect}
\title{Connect Duplicated Values}
\usage{
dup.connect(data, id, dup.var)
}
\arguments{
\item{data}{dataframe or matrix}

\item{id}{id column names or indexs}

\item{dup.var}{duplicated column names or indexs}
}
\value{
dataframe contains id and duplicated values
}
\description{
Connect Duplicated Values
}
\examples{
dup.connect(data = mtcars,id = 'am',dup.var = 'cyl')
dup.connect(data = mtcars,
             id = c('am','gear'),
             dup.var = c('cyl','qsec'))
}
