% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate.R
\name{duplicate}
\alias{duplicate}
\title{find all duplicated ones}
\usage{
duplicate(x, column_between = FALSE, row_between = FALSE,
  row_in = FALSE, column_in = FALSE, every = FALSE)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix}

\item{column_between}{TRUE is to detect the duplicated rows.}

\item{row_between}{TRUE is to detect the duplicated columns.}

\item{row_in}{TRUE is to detect the duplicated values in each row.}

\item{column_in}{TRUE is to detect the duplicated values in each column.}

\item{every}{TRUE is to detect the duplicated values in dataframe or matrix.}
}
\value{
TRUE as duplicated, FALSE as not
}
\description{
find all duplicated ones
}
\examples{
x <- c('a','b','a',1,1,2)
duplicate(x)
a = data.frame(k=c(1,1,2,1,4),
l=c(1,1,'a',5,6))
duplicate(a,row_in = TRUE)
duplicate(a,row_between = TRUE)
duplicate(a,column_between = TRUE)
duplicate(a,column_in = TRUE)
}
