% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{duplicated_all}
\alias{duplicated_all}
\title{Determine All Duplicate Elements}
\usage{
duplicated_all(x)
}
\arguments{
\item{x}{character}
}
\value{
logical value
}
\description{
Determine All Duplicate Elements
}
\examples{
x=c(1,3,2,1,2)
duplicated(x)
duplicated_all(x)
}
