% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{duplicated_last}
\alias{duplicated_last}
\title{Determine Duplicate Elements in the Last Position}
\usage{
duplicated_last(x)
}
\arguments{
\item{x}{character}
}
\value{
logical value
}
\description{
Determine Duplicate Elements in the Last Position
}
\examples{
x=c(1,3,2,1,2)
duplicated(x)
duplicated_last(x)
}
