% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_length.R
\name{equal_length}
\alias{equal_length}
\title{Equal Length}
\usage{
equal_length(x, suffix = " ", nchar, colname = FALSE, rowname = FALSE)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix.}

\item{suffix}{suffix}

\item{nchar}{maximun length}

\item{colname}{a logistic value, default is FALSE}

\item{rowname}{a logistic value, default is FALSE}
}
\value{
equal length results
}
\description{
Equal Length
}
\examples{
a=c(123,1,24,5,1.22554)
equal_length(a,0)

df = data.frame(
    a=c(12,1,1.23),
    b=c('a','abcd','d')
)
equal_length(x = df,suffix = 'x')

equal_length(x = df,suffix = 0,nchar =5)
}
