% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec}
\alias{exec}
\title{execute string command
This command just execute in the paraent frame.}
\usage{
exec(string, envir = parent.frame())
}
\arguments{
\item{string}{one string}

\item{envir}{the environment in which sting is to be evaluated.}
}
\value{
execute string command
}
\description{
execute string command
This command just execute in the paraent frame.
}
\examples{
a=2
exec('a = 1')
a
}
