% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand Data by Weight}
\usage{
expand(x, weight)
}
\arguments{
\item{x}{dataframe or matrix}

\item{weight}{weight column names or index}
}
\value{
expanded data
}
\description{
Expand Data by Weight
}
\examples{
df=data.frame(v=c(1,2,3),
               x=c(7,8,9),
               n=c(2,3,4))
expand(x = df,weight = 3)
expand(x = df,weight = 'n')
}
