% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{factor.nms}
\alias{factor.nms}
\title{Return factor names in matrix or dataframe}
\usage{
factor.nms(df)
}
\arguments{
\item{df}{dataframe or matrix}
}
\value{
factor names vectors
}
\description{
Return factor names in matrix or dataframe
}
