% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.dir.R
\name{file.dir}
\alias{file.dir}
\title{up level directory}
\usage{
file.dir(path, end.slash = TRUE, extension = TRUE)
}
\arguments{
\item{path}{path of file}

\item{end.slash}{logical. Whether to end with slash}

\item{extension}{logical. whether file name include extension}
}
\value{
upper directory
}
\description{
up level directory
}
