% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.name.R
\name{file.name}
\alias{file.name}
\title{Extract file name}
\usage{
file.name(..., extension = TRUE)
}
\arguments{
\item{...}{one or more file path}

\item{extension}{whether include extension, default is TRUE}
}
\value{
file names
}
\description{
Extract file name
}
\examples{
file.name('f:/dir/1.txt')
file.name('f:/dir/1.txt', 'f:/dir/1.txt')
file.name('f:/dir/1.txt', 'f:/dir/1.txt', 'f:/dir/')
}
