% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.dir.R
\name{formal_dir}
\alias{formal_dir}
\title{formal directory}
\usage{
formal_dir(dir, end.slash = FALSE)
}
\arguments{
\item{dir}{one directory}

\item{end.slash}{logical}
}
\value{
formed directory
}
\description{
formal directory
}
