% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get Names of Object}
\usage{
get_names(...)
}
\arguments{
\item{...}{any type of data object}
}
\value{
names of object
}
\description{
Return the names of input. For example: if you input a, you will get 'a'.
}
\examples{
a=c(1,2,3)
get_names(a,mtcars)
}
