% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_names.R
\name{give_names}
\alias{give_names}
\alias{give_names.character}
\alias{give_names.numeric}
\alias{give_names.logical}
\alias{give_names.list}
\alias{give_names.data.frame}
\alias{give_names.matrix}
\title{change vector, dataframe or matrix names}
\usage{
give_names(data, ...)

\method{give_names}{character}(data, ...)

\method{give_names}{numeric}(data, ...)

\method{give_names}{logical}(data, ...)

\method{give_names}{list}(data, ...)

\method{give_names}{data.frame}(data, ..., row = FALSE)

\method{give_names}{matrix}(data, ..., row = FALSE)
}
\arguments{
\item{data}{one vector, list, dataframe or matrix}

\item{...}{one or more names}

\item{row}{logical, whether the names is row names. Default is FALSE}
}
\value{
names changed data
}
\description{
change vector, dataframe or matrix names
}
