% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes_equals_grapes.R
\name{\%==\%}
\alias{\%==\%}
\title{Locate Accurately}
\usage{
a \%==\% b
}
\arguments{
\item{a}{vector for matching}

\item{b}{vector for searching}
}
\value{
If length of a is one, a vector will be return. If length of a is more
    than one, a list for each element will be return.
}
\description{
Locate Accurately
}
\examples{
a=c(1,2,3,4)
b=c(1,2,3,1,4,1,5,6,1,4,1)
a \%==\% b
}
