% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate vectors after converting to character.}
\usage{
a \%+\% b
}
\arguments{
\item{a}{one R objects, to be converted to character vectors.}

\item{b}{one R objects, to be converted to character vectors.}
}
\value{
one vector
}
\description{
Concatenate vectors after converting to character.
}
\examples{
1 \%+\% 1
}
