% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes_similar_equals_grapes.R
\name{\%s=\%}
\alias{\%s=\%}
\title{Locate Similarly by grep()}
\usage{
a \%s=\% b
}
\arguments{
\item{a}{vector for matching}

\item{b}{vector for searching}
}
\value{
A list contains location information.
}
\description{
Locate Similarly by grep()
}
\examples{
1 \%s=\% c(1,12,3)
c(1,2) \%s=\% c(1,12,3)
}
