% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grepl.R
\name{Grepl}
\alias{Grepl}
\title{Judge for Included Character}
\usage{
Grepl(pattern, x)
}
\arguments{
\item{pattern}{one or more vectors}

\item{x}{one or more vectors}
}
\value{
a matrix with logical words
}
\description{
Judge for Included Character
}
\details{
'%or%' and '%and%' can be used in pattern, see examples.
}
\examples{

a=c('abcd','agj','abcu')

# Grepl for one vector
pat1='b'
Grepl(pat1,a)

# Grepl for two vectors
pat2=c('c','d')
Grepl(pat2,a)

# use \%or\% in pattern
pat3=c('a\%or\%c','d')
Grepl(pat3,a)

# use \%and\% in pattern
pat4=c('a\%and\%c','d')
Grepl(pat4,a)
}
