% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_node.R
\name{has_children}
\alias{has_children}
\title{Wether children nodes exist}
\usage{
has_children(...)
}
\arguments{
\item{...}{one or more documents, nodes, or node sets.}
}
\value{
logical value
}
\description{
Wether children nodes exist
}
\examples{
txt='<div class="activityBody postBody thing">
        <p>
            <a href="/forum/conversation" rel="post" >(22)</a>
            where?
        </p>
      <p>
        stays 
        <b>disappears</b>
        <a>disappears</a>D
        <span>disappears</span>
        stays
      </p>
    </div>'
 library(xml2)
 html=read_html(txt)
 has_children(html)
}
