% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_decrease.R
\name{increase}
\alias{increase}
\title{increase character}
\usage{
increase(chr)
}
\arguments{
\item{chr}{one vector}
}
\value{
increased vector
}
\description{
increase character
}
\examples{
set.seed(2020)
x=rnorm(20)
increase(x)
}
