% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_Add_Symbol.R
\name{inner_Add_Symbol}
\alias{inner_Add_Symbol}
\title{Concatenate Strings}
\usage{
inner_Add_Symbol(x, symbol = "+")
}
\arguments{
\item{x}{vectors}

\item{symbol}{defulat is '+'}
}
\value{
a concatenated string
}
\description{
Concatenate vectors by adding a symbol.
}
\examples{
inner_Add_Symbol(c('a','b'))
inner_Add_Symbol(c('a','b'),"$")
inner_Add_Symbol(c('a','b'),"")
}
