% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Rversion.R
\name{install_Rversion}
\alias{install_Rversion}
\title{Install contributed packages by R version}
\usage{
install_Rversion(..., platform, Rversion = NULL, lib = ".")
}
\arguments{
\item{...}{one or more package}

\item{platform}{windows or mac}

\item{Rversion}{version of R}

\item{lib}{path}
}
\value{
contributed packages
}
\description{
Install contributed packages by R version
}
