% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Keep objects}
\usage{
keep(..., envir = .GlobalEnv)
}
\arguments{
\item{...}{one or more objects}

\item{envir}{environment, default is global}
}
\description{
Keep objects
}
\examples{
\donttest{
a <- 1
b <- 2
d <- 4
keep(a)
}
}
