% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knife.R
\name{knife}
\alias{knife}
\alias{knife_left}
\alias{knife_right}
\title{Knife characters}
\usage{
knife_left(x, n)

knife_right(x, n)
}
\arguments{
\item{x}{one character}

\item{n}{number}
}
\description{
Knife characters
}
\examples{
knife_left(123,2)
knife_right(123,2)
}
