% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Select character from last}
\usage{
last(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{If missing, the last element will be used.}
}
\value{
last element
}
\description{
Select character from last
}
\examples{
letters |> last()
letters |> last(1:2)
}
