% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last_column}
\alias{last_column}
\title{Select dataframe column from last}
\usage{
last_column(x, n)
}
\arguments{
\item{x}{dataframe}

\item{n}{If missing, the last element will be used.}
}
\value{
last column
}
\description{
Select dataframe column from last
}
\examples{
mtcars |> last_column()
mtcars |> last_column(1:2)
}
