% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last_row}
\alias{last_row}
\title{Select dataframe row from last}
\usage{
last_row(x, n)
}
\arguments{
\item{x}{dataframe}

\item{n}{If missing, the last element will be used.}
}
\value{
last row
}
\description{
Select dataframe row from last
}
\examples{
mtcars |> last_row()
mtcars |> last_row(1:2)
}
