% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{Truncate Characters from the Left}
\usage{
left(x, n)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix.}

\item{n}{length}
}
\value{
substring
}
\description{
Truncate Characters from the Left
}
\examples{
left("abcd",3)
left(c("abc","gjh"),2)
df = data.frame(
  a = c(123,234,456),
  b = c("abc","bcd","hjg")
)
left(df,2)
}
