% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_equal.R
\name{left_equal}
\alias{left_equal}
\title{Compare two characters from left
Much useful for arguments input. Case is ignored.}
\usage{
left_equal(x1, x2)
}
\arguments{
\item{x1}{one character}

\item{x2}{the other character}
}
\value{
logical
}
\description{
Compare two characters from left
Much useful for arguments input. Case is ignored.
}
\examples{
left_equal('o','OK')
left_equal('ok','O')
left_equal('ok','Ok')
}
