% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extdata.R
\name{load_extdata}
\alias{load_extdata}
\title{Load external data from R package}
\usage{
load_extdata(package, file)
}
\arguments{
\item{package}{one package name}

\item{file}{one file name}
}
\value{
path of data
}
\description{
Load external data from R package
}
