% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{mid}
\alias{mid}
\title{Truncate Characters from the Inside}
\usage{
mid(x, start, n = 1e+11)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix.}

\item{start}{starting position}

\item{n}{length, n can be less than zero}
}
\value{
substring
}
\description{
Truncate Characters from the Inside
}
\examples{
mid("abcd",3,1)
mid(c("abc","gjh"),2,2)
df = data.frame(
  a = c(123,234,456),
  b = c("abc","bcd","hjg")
)
mid(df,2,1)
mid(df,2,-2)
}
