% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror.current}
\alias{mirror.current}
\title{Current mirror}
\usage{
mirror.current()
}
\value{
a list contains CRAN and Bioconductor mirror
}
\description{
Current mirrors of CRAN adn Bioconductor
}
\examples{
mirror.current()
}
