% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror.speed}
\alias{mirror.speed}
\title{Test speed of mirror}
\usage{
mirror.speed(min.second = 0.2, cran = TRUE, bioc = TRUE)
}
\arguments{
\item{min.second}{the minium second time to visit the mirror web page}

\item{cran}{logical, whether to test CRAN mirrors. Default is TRUE}

\item{bioc}{logical, whether to test bioconductor mirrors. Default is TRUE}
}
\value{
repositories which visiting time is minus the minium second.
}
\description{
Test speed of mirror
}
