% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.data.R
\name{model.data}
\alias{model.data}
\title{Extract data of model}
\usage{
model.data(fit)
}
\arguments{
\item{fit}{fitted results}
}
\value{
dataframe in the model
}
\description{
Extract data of model
}
\examples{
fit <- lm(mpg~vs+am+poly(qsec,2),data=mtcars)
head(model.data(fit))
}
