% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAcount.R
\name{NA.col.sums}
\alias{NA.col.sums}
\title{Sum of missing value by column}
\usage{
NA.col.sums(data)
}
\arguments{
\item{data}{must be dataframe or matrix}
}
\value{
sum of missing value by column
}
\description{
NA is treated as missing value.
}
\examples{
df = data.frame(x=rep(c(1,NA,2,NA,6,NA),10),
                 y=rep(c(1,NA,2),20))
NA.col.sums(df)
}
