% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAcount.R
\name{NA.whole.sums}
\alias{NA.whole.sums}
\title{Sum of missing value in the whole dataframe}
\usage{
NA.whole.sums(data)
}
\arguments{
\item{data}{must be dataframe or matrix}
}
\value{
sum of missing value in the whole dataframe
}
\description{
NA is treated as missing value.
}
\examples{
df = data.frame(x=rep(c(1,NA,2,NA,6,NA),10),
                 y=rep(c(1,NA,2),20))
NA.whole.sums(df)
}
