% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{name.n}
\alias{name.n}
\title{Names with different letters}
\usage{
name.n(df, least = NULL, most = NULL)
}
\arguments{
\item{df}{datafame or matrix}

\item{least}{names with at least different letters, which means >=}

\item{most}{names with at most different letters, which means <=}
}
\value{
names
}
\description{
Names with different letters
}
