% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{names_n}
\alias{names_n}
\title{Names with different letters}
\usage{
names_n(df, most = NULL, least = NULL)
}
\arguments{
\item{df}{datafame or matrix}

\item{most}{names with at most different letters, which means <=}

\item{least}{names with at least different letters, which means >=}
}
\value{
names
}
\description{
Names with different letters
}
