% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nchar.R
\name{Nchar}
\alias{Nchar}
\title{Number of Characters}
\usage{
Nchar(x)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix.}
}
\value{
number of characters in each location
}
\description{
Number of Characters
}
\examples{
Nchar("abcd")
Nchar(c("abc","gjh"))
df = data.frame(
  a = c(1,12,12.3),
  b = c("a","ab","abc")
)
Nchar(df)
}
