% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{numeric.nms}
\alias{numeric.nms}
\title{Return numeric names in matrix or dataframe}
\usage{
numeric.nms(df)
}
\arguments{
\item{df}{dataframe or matrix}
}
\value{
numeric names vectors
}
\description{
Return numeric names in matrix or dataframe
}
