% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_all.R
\name{pakcage_all}
\alias{pakcage_all}
\title{Get all functions in one package}
\usage{
pakcage_all(x)
}
\arguments{
\item{x}{package}
}
\value{
all functions in one package
}
\description{
Get all functions in one package
}
