% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste0_columns.R
\name{paste0_columns}
\alias{paste0_columns}
\title{Paste Columns Together}
\usage{
paste0_columns(df, collapse = ",")
}
\arguments{
\item{df}{a dataframe}

\item{collapse}{collapse, default is comma}
}
\value{
a character
}
\description{
Paste each column in a dataframe together.
}
\examples{
df=data.frame(a=c(1,2,30),
              b=c('x','y','z'))
paste0_columns(df)

df=data.frame(a=c(1,2,30),b=c('x','y','z'),c=c(1,7,8))
paste0_columns(df)

}
