% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2df.R
\name{rd2df}
\alias{rd2df}
\title{Convert package Rd file under man directory into dataframe}
\usage{
rd2df(pkg)
}
\arguments{
\item{pkg}{source package path unzip from "tar.gz" file}
}
\value{
one dataframe
}
\description{
Convert package Rd file under man directory into dataframe
}
