% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_R.R
\name{read_R}
\alias{read_R}
\title{Read R file}
\usage{
read_R(R, pattern)
}
\arguments{
\item{R}{path of R file}

\item{pattern}{pattern}
}
\value{
one vector of R command with names of R file
}
\description{
Read R file
}
