% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace.R
\name{Replace}
\alias{Replace}
\title{Replace}
\usage{
Replace(data, from, to, pattern, ignore.case = FALSE)
}
\arguments{
\item{data}{can be number, strings, verctors, dataframe or matrix.}

\item{from}{replaced stings}

\item{to}{replacements}

\item{pattern}{like from:to}

\item{ignore.case}{logical, whether to ignore case}
}
\value{
replaced data
}
\description{
There are two methods in this function. You can use repalce many objects to one by form and to. pattern can be used to one object replaced by the other one.
}
\examples{
Replace(data = 232,from = 2,to = 1)
Replace(data = c(232,'a4b'),
        from = c(2,'.*4'),to = 1,
        pattern = c('a:e','b:h'))
df = data.frame(
  a = c(232, 452),
  b = c("nba", "cba")
)
Replace(data = df,
        from = 2,to = 1,
        pattern = c('a:e','b:h'))

}
