% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace0.R
\name{Replace0}
\alias{Replace0}
\title{Replaced by Empty}
\usage{
Replace0(data, from)
}
\arguments{
\item{data}{can be number, strings, verctors, dataframe or matrix.}

\item{from}{replaced stings}
}
\value{
replaced data
}
\description{
Replaced by Empty
}
\examples{
Replace0(data = 232,from = 2)
Replace0(data = c(232,'a4b'),from = c(2,'.*4'))

df = data.frame(
  a = c(232, 452),
  b = c("nba", "cba")
)
Replace0(data = df, from = c(2,'a'))

}
