% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_ex.R
\name{Replace_ex}
\alias{Replace_ex}
\title{Replace Exactly}
\usage{
Replace_ex(x, from, to, pattern)
}
\arguments{
\item{x}{vector, dataframe or matrix}

\item{from}{replaced stings}

\item{to}{replacements}

\item{pattern}{a special pattern, see examples for detail}
}
\value{
replaced data
}
\description{
Replace Exactly
}
\examples{
a=c(1,2,3,1,4)
Replace_ex(x = a,from = c(1,2),to=5)
Replace_ex(x=a,pattern = c('1:5','2:5'))
Replace_ex(x=a,pattern = '[12]:5')


a=data.frame(v=c(1,2,3,2,4),
             b=c(7,8,9,4,6))
Replace_ex(x = a,from = c(1,2),to=5)
Replace_ex(x=a,pattern = c('1:5','2:5'))
}
