% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{replicate}
\alias{replicate}
\alias{rep_n}
\alias{rep_character}
\title{Replicate Each Elements of Vectors}
\usage{
rep_n(x, each)

rep_character(x, each)
}
\arguments{
\item{x}{vectors}

\item{each}{one or more numbers for replication}
}
\value{
replicated vectors
}
\description{
Replicate Each Elements of Vectors
}
\examples{
rep_n(c('ab','cde','k','op'),5)
rep_n(c('ab','cde','k','op'),c(4,6))
rep_n(c('ab','cde','k','op'),c(1,2,3,4))

rep_character(c('ab','cde','k','op'),5)
rep_character(c('ab','cde','k','op'),c(4,6))
rep_character(c('ab','cde','k','op'),c(1,2,3,4))
}
