% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_toLong.R
\name{reshape_toLong}
\alias{reshape_toLong}
\title{Convert Wide Data to Long}
\usage{
reshape_toLong(data, prefix = NULL, suffix = NULL, var.names = NULL)
}
\arguments{
\item{data}{wide data}

\item{prefix}{prefix of value variables}

\item{suffix}{suffix of value variables}

\item{var.names}{names of value variables, do.value will be created as the name of value column}
}
\value{
long data
}
\description{
It is easy to convert wide data to long in this function. Be careful, id must be unique. prefix, suffix and var.names can be used together.
}
