% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Truncate Characters from the Right}
\usage{
right(x, n)
}
\arguments{
\item{x}{can be number, strings, verctors, dataframe or matrix.}

\item{n}{length}
}
\value{
substring
}
\description{
Truncate Characters from the Right
}
\examples{
right("abcd",3)
right(c("abc","gjh"),2)
df = data.frame(
  a = c(123,234,456),
  b = c("abc","bcd","hjg")
)
right(df,2)
}
