% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_equal.R
\name{right_equal}
\alias{right_equal}
\title{Compare two characters from right
Much useful for arguments input. Case is ignored.}
\usage{
right_equal(x1, x2)
}
\arguments{
\item{x1}{one character}

\item{x2}{the other character}
}
\value{
logical
}
\description{
Compare two characters from right
Much useful for arguments input. Case is ignored.
}
\examples{
right_equal('k','OK')
right_equal('ok','k')
right_equal('ok','Ok')
}
