% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_nchar.R
\name{rm_nchar}
\alias{rm_nchar}
\title{Remove elements by number of characters}
\usage{
rm_nchar(x, least, most)
}
\arguments{
\item{x}{one vector}

\item{least}{least number of characters}

\item{most}{most number of characters}
}
\value{
removed vector
}
\description{
Remove elements by number of characters
}
\examples{
x <- c('a','abc','abcd',NA)
rm_nchar(x,least = 1)
rm_nchar(x,most = 4)
rm_nchar(x,least = 1, most = 4)
}
