% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.freq.R
\name{row.freq}
\alias{row.freq}
\title{Row Frequency}
\usage{
row.freq(x)
}
\arguments{
\item{x}{dataframe or matrix}
}
\value{
data with frequency column
}
\description{
Row Frequency
}
\examples{
row.freq(x=mtcars[,8:11])
}
