% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{seq_range}
\alias{seq_range}
\title{sequence range of one vector}
\usage{
seq_range(x, by = 1)
}
\arguments{
\item{x}{one vector}

\item{by}{default is 1}
}
\value{
number sequence
}
\description{
sequence range of one vector
}
\examples{
seq_range(letters)
seq_range(letters,2)
}
