% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{set_Bioc_mirror}
\alias{set_Bioc_mirror}
\title{set bioconductor mirror}
\usage{
set_Bioc_mirror(url)
}
\arguments{
\item{url}{mirror url}
}
\value{
set bioconductor mirror
}
\description{
set bioconductor mirror
}
