% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_function.R
\name{show_function}
\alias{show_function}
\title{Show function command line in new script
script will be store in your temporary directory}
\usage{
show_function(f, file = NULL)
}
\arguments{
\item{f}{one function}

\item{file}{file name}
}
\value{
command line in new script
}
\description{
Show function command line in new script
script will be store in your temporary directory
}
