% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_expand.R
\name{split_expand}
\alias{split_expand}
\title{Split One Column and Expand}
\usage{
split_expand(data, variable, sep)
}
\arguments{
\item{data}{dataframe or matrix}

\item{variable}{one column name with connected values}

\item{sep}{seperated symbol, which can be one or more}
}
\value{
expanded dataframe or matrix
}
\description{
Split One Column and Expand
}
\examples{
df=data.frame(a=c(1,0),
              b=c('a','n'),
           cyl=c('6;6;4;4;4',
                 '6;8;'))
split_expand(data=df,variable='cyl',sep=';')
}
