% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Target_download_dir.R,
%   R/Target_download_file.R, R/Target_download_url.R
\name{Target}
\alias{Target}
\alias{Target_download_dir}
\alias{Target_download_file}
\alias{Target_download_url}
\title{Download data from Target database}
\usage{
Target_download_dir(url)

Target_download_file(url)

Target_download_url(url)
}
\arguments{
\item{url}{url link of data}
}
\value{
invisible url and downloaded data
}
\description{
Download data from Target database
}
\examples{
\donttest{
Target_download_dir('https://target-data.nci.nih.gov/Public/ALL/clinical/')
}
\donttest{
# download data by file url
Target_download_dir("https://target-data.nci.nih.gov/Public/OS/Brazil/gene_expression_array/METADATA/MANIFEST.txt")
}
\donttest{
# get file url
Target_download_dir('https://target-data.nci.nih.gov/Public/ALL/clinical/')
}
}
