% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim.R
\name{Trim}
\alias{Trim}
\alias{Trim_left}
\alias{trim-left}
\alias{Trim_right}
\alias{trim-right}
\title{Trim}
\usage{
Trim(x, pattern = " ")

Trim_left(x, pattern = " ")

Trim_right(x, pattern = " ")
}
\arguments{
\item{x}{can be vector or dataframe or matrix}

\item{pattern}{one or more pattern pattern}
}
\value{
a trimed string
}
\description{
Trim
}
