% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.no.NA.R
\name{unique_no.NA}
\alias{unique_no.NA}
\title{Unique Without NA}
\usage{
unique_no.NA(x)
}
\arguments{
\item{x}{vector}
}
\value{
unique values with no NA
}
\description{
Unique Without NA
}
\examples{
x=c(1,2,3,1,NA)
unique(x)
unique_no.NA(x)
}
