% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_rm_all.R
\name{unlibray}
\alias{unlibray}
\title{Detach package}
\usage{
unlibray(x)
}
\arguments{
\item{x}{one package name, if missing, detach all packages}
}
\value{
detach one package
}
\description{
Detach package
}
