% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper.dir.R
\name{upper.dir}
\alias{upper.dir}
\title{up level directory}
\usage{
upper.dir(dir, end.slash = TRUE)
}
\arguments{
\item{dir}{present directory or file path}

\item{end.slash}{logical}
}
\value{
upper directory
}
\description{
up level directory
}
