% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write data to Excel file
write or append one or more data into one Excel file in each sheet.}
\usage{
write_xlsx(
  ...,
  file,
  sheet,
  col.names = TRUE,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{...}{one or more data}

\item{file}{Excel file name}

\item{sheet}{sheet names}

\item{col.names}{logical, whether to write out column names}

\item{row.names}{logical, whether to write out row names}

\item{overwrite}{logical, whether to overwrite an existing file}

\item{append}{logical, whether to add data to an existing file}
}
\value{
write one or more data into one Excel file
}
\description{
Write data to Excel file
write or append one or more data into one Excel file in each sheet.
}
\examples{
\donttest{
mtcars2 = mtcars
# write_xlsx(mtcars,mtcars2,file='mtcars')
}
}
