% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_split.R
\name{col_split}
\alias{col_split}
\title{Split A Vector into Columns}
\usage{
col_split(x, split, reg_expr, colnames)
}
\arguments{
\item{x}{a vector}

\item{split}{character. Split exactly}

\item{reg_expr}{character. Split by regular expressions}

\item{colnames}{optional. Column names for outcome}
}
\value{
A dataframe with several columns.
}
\description{
Split A Vector into Columns
}
\examples{
x=c('1a2','3a4','4a4')
col_split(x,split='a')
col_split(x = x,reg_expr = '[a-z]')
}
