% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\title{Find Common Objects from Vectors}
\usage{
common(x)
}
\arguments{
\item{x}{must be list}
}
\value{
common objects
}
\description{
Find Common Objects from Vectors
}
\examples{

x1=c('a','e','d')
x2=c('a','c','e')
x3=c('a','e','j','d')
x=list(x1,x2,x3)
common(x)

}
