% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_toWide.R
\name{reshape_toWide}
\alias{reshape_toWide}
\title{Reshape to Wide Format}
\usage{
reshape_toWide(
  data,
  key = NULL,
  value = NULL,
  prefix = NULL,
  suffix = NULL,
  sep = "_"
)
}
\arguments{
\item{data}{long data}

\item{key}{column names for key, which can be one or more}

\item{value}{column names for exchange, which can be one or more}

\item{prefix}{column names for prefix, which can be one or more}

\item{suffix}{column names for suffix, which can be one or more}

\item{sep}{seperation}
}
\value{
A wide data.
}
\description{
Reshape to Wide Format
}
